feat: refactor risk optimization into smaller units
